# wait for leader to transfer back during rolling restart