--- a/pkg/cluster/api/pdapi.go
+++ b/pkg/cluster/api/pdapi.go
@@ -539,15 +539,40 @@ func (pc *PDClient) EvictStoreLeader(host string, retryOpt *utils.RetryOption, c
 // RecoverStoreLeader waits for some leaders to transfer back.
 //
 // Currently, recoverStoreLeader will be considered as succeed in any of the following case
-//   - 2/3 of leaders are already transferred back.
-//   - The leader count has been unchanged for 5 times.
-//     The default leadership transfer timeout for one region is 10s,
-//     so set the default value to about 10s (5*2).
+//
+//  1. 2/3 of leaders are already transferred back.
+//
+//  2. Original leader count is less than 60.
+//
+//     Since PD considers it as balance when the leader count delta is less than 10, so
+//     these two conditions should be taken into consideration
+//
+//     - When the original leader count is less than 20, there is possibility that
+//     no leader will transfer back.
+//     For example: The target store's leader count is 19. Other stores' leader count are 9.
+//     There are 20 stores in total. In this case, there may be no leader to transfer back.
+//
+//     - When the leader count is less than 60, there is possibility that only less than 2/3
+//     leaders are transfered back. `(N-10-10 >= 2/3*N) -> (N>=60)`.
+//     For example: The target store's leader count is 59. Other stores' leader count are 49.
+//     There are 50 stores in total. In this case, there may be only 39 leaders to transfer back.
+//
+//  3. The leader count has been unchanged for 5 times.
 func (pc *PDClient) RecoverStoreLeader(host string, originalCount int, retryOpt *utils.RetryOption, countLeader func(string) (int, error)) error {
+	// When the leader count is less than 60, just ignore recovering.
+	if originalCount <= 60 {
+		return nil
+	}
+
 	targetCount := originalCount * 2 / 3
+	// The default leadership transfer timeout for one region is 10s,
+	// so set the default value to about 10s (5*2s=10s).
+	// NOTE: PD may not transfer leader to a newly started store in the future,
+	// (check https://github.com/tikv/pd/pull/4762 for details),
+	// so this strategy should also be enhanced later.
 	maxUnchangedTimes := 5
 
-	// get info of current stores
+	// Get info of current stores.
 	latestStore, err := pc.GetCurrentStore(host)
 	if err != nil {
 		if errors.Is(err, ErrNoStore) {
@@ -558,11 +583,14 @@ func (pc *PDClient) RecoverStoreLeader(host string, originalCount int, retryOpt 
 
 	pc.l().Infof("\tRecovering about %d leaders to store %s, original count is %d...", targetCount, latestStore.Store.Address, originalCount)
 
-	// wait for the transfer to complete
+	// Wait for the transfer to complete.
 	if retryOpt == nil {
 		retryOpt = &utils.RetryOption{
+			// The default timeout of evicting leader is 600s, so set the recovering timeout to
+			// 2/3 of it should be reasonable. Besides, One local test shows it takes about
+			// 30s to recover 3.6k leaders.
+			Timeout: time.Second * 400,
 			Delay:   time.Second * 2,
-			Timeout: time.Second * 600,
 		}
 	}
 
@@ -587,7 +615,7 @@ func (pc *PDClient) RecoverStoreLeader(host string, originalCount int, retryOpt 
 			return nil
 		}
 
-		// Check if the leader count has been unchanged for too much times.
+		// Check if the leader count has been unchanged for certain times.
 		if lastLeaderCount == curLeaderCount {
 			curUnchangedTimes += 1
 			if curUnchangedTimes >= maxUnchangedTimes {
@@ -604,7 +632,7 @@ func (pc *PDClient) RecoverStoreLeader(host string, originalCount int, retryOpt 
 			targetCount-curLeaderCount,
 		)
 
-		// return error by default, to make the retry work
+		// Return error by default, to make the retry work.
 		return perrs.New("still waiting for the store leaders to transfer back")
 	}, *retryOpt); err != nil {
 		return fmt.Errorf("error recovering store leader to %s, %v", host, err)