--- a/pkg/cluster/api/pdapi.go
+++ b/pkg/cluster/api/pdapi.go
@@ -542,7 +542,11 @@ func (pc *PDClient) EvictStoreLeader(host string, retryOpt *utils.RetryOption, c
 //
 //  1. 2/3 of leaders are already transferred back.
 //
-//  2. Original leader count is less than 57.
+//  2. Original leader count is less than 200.
+//     Though the accurate threshold is 57, it can be set to a larger value, for example 200.
+//     Moreover, clusters which have small number of leaders are supposed to has low pressure,
+//     and this recovering strategy may be unnecessary for them. Clusters in production env
+//     usually has thousands of leaders.
 //
 //     Since PD considers it as balance when the leader count delta is less than 10, so
 //     these two conditions should be taken into consideration
@@ -555,12 +559,14 @@ func (pc *PDClient) EvictStoreLeader(host string, retryOpt *utils.RetryOption, c
 //     - When the leader count is less than 57, there is possibility that only less than 2/3
 //     leaders are transfered back. `(N-10-9 >= 2/3*N) -> (N>=57)`.
 //     For example: The target store's leader count is 56. Other stores' leader count are 46.
-//     There are 57 stores in total. In this case, there may be only 37 leaders to transfer back.
+//     There are 57 stores in total. In this case, there may be only 37 leaders to transfer back,
+//     and 37/56 < 2/3. Accordingly, if the target store's leader count is 57, then there may be
+//     38 leaders to transfer back, and 38/57 == 2/3.
 //
 //  3. The leader count has been unchanged for 5 times.
 func (pc *PDClient) RecoverStoreLeader(host string, originalCount int, retryOpt *utils.RetryOption, countLeader func(string) (int, error)) error {
 	// When the leader count is less than certain number, just ignore recovering.
-	if originalCount < 57 {
+	if originalCount < 200 {
 		return nil
 	}