@cosven 2022-10-09
I'll add another rule later: when the leader count is less than 50, just ignore recovering process, because PD does not balance leader when two tikv instances's leader count diff is less than 10.